refactor: move `ITEMS_PER_PAGE` to config file
